requirements.txt
gymnasium
numpy
pygame
setuptools
pytest
torch.utils
tensorboard
black
flake8
pylint
pyyaml
